reduce container component size
