+++ app/javascript/react/components/FileUpload/UploadSelect/UploadType.jsx
