+++ app/javascript/react/components/FileUpload/index.js
@@ -1,9 +1,8 @@
 export {default as FailedUrlList} from './FailedUrlList/FailedUrlList';
 export {default as FileList} from './FileList/FileList';
 export {TabularCheckStatus} from './FileList/File';
-export {default as Instructions} from './Instructions';
 export {default as ModalUrl} from './ModalUrl';
 export {default as ModalValidationReport} from './ModalValidationReport/ModalValidationReport';
-export {default as UploadType} from './UploadType';
+export {default as UploadSelect} from './UploadSelect/UploadSelect';
 export {default as ValidateFiles} from './ValidateFiles';
 export {default as WarningMessage} from './WarningMessage';
