+++ app/views/stash_engine/resources/upload.html.erb
@@ -11,7 +11,7 @@
 <% end %>
 
 <!-- map(&:attributes) one way for translating ActiveRecord field type to json at React side adequately -->
-<%= react_component('containers/FileUpload/UploadFiles', {
+<%= react_component('containers/UploadFiles', {
   resource_id: @resource.id,
   readme_size: @readme_size,
   file_uploads: @resource.generic_files.validated_table.as_json(
